splitting packages, added packages interface, moving yaml loading to interface
